+++ packages/core/src/modules/mark/mark-error.ts
@@ -1,5 +1,5 @@
 import { ApplicationError } from "../../error"
-import { PROBLEM_DETAILS } from "../../problem-details-registry"
+import { PROBLEM_DETAILS } from "../../http-problem-details"
 
 export class MarkNotFoundError extends ApplicationError {
   constructor(id: string) {
