+++ packages/core/src/modules/payment/payment-error.ts
@@ -1,5 +1,5 @@
 import { ApplicationError } from "../../error"
-import { PROBLEM_DETAILS } from "../../problem-details-registry"
+import { PROBLEM_DETAILS } from "../../http-problem-details"
 
 export class StripeAccountNotFoundError extends ApplicationError {
   constructor(id: string) {
